update cli and readme
