+++ README.md
@@ -20,9 +20,7 @@ The Installation if ffmpeg differ among operating systems. If you use [Anaconda]
 conda install -c conda-forge ffmpeg
 ```
 
-__Decoding__ is supported with any recent build of ffmpeg. For __Encoding__ it is recommended to use the Fraunhofer AAC encoder (`libfdk_aac`) which is not included in the default ffmpeg builds. Note that the conda version currently does _not_ include `fdk-aac`. If `libfdk_aac` is not installed _stempeg_ will use the default `aac` codec which will result in slightly inferior audio quality.
-
-You can install ffmpeg with `libfdk-aac` support manually as following:
+Note that for better quality encoding it is recommended to ffmpeg with `libfdk-aac` support manually as following:
 
 * Mac: use homebrew: `brew install ffmpeg --with-fdk-aac`
 * Ubuntu Linux: See installation script [here](https://gist.github.com/rafaelbiriba/7f2d7c6f6c3d6ae2a5cb).
@@ -30,17 +28,21 @@ You can install ffmpeg with `libfdk-aac` support manually as following:
 
 ### 2. Installation of the _stempeg_ package
 
-Installation via PyPI using pip
+A) Installation via PyPI using pip
 
 ```
 pip install stempeg
 ```
 
+B) Installation via conda
+
+```
+conda install -c conda-forge stempeg
+```
+
 ## Usage
 
-There are very few freely available stem files. We included a small test track from the Canadian rock-band _The Easton Ellises_. The band [released them](https://www.heise.de/ct/artikel/c-t-Remix-Wettbewerb-The-Easton-Ellises-2542427.html) under Creative Commons license CC BY-NC-SA 3.0.
-
-To use the included stem example you can use `stempeg.example_stem_path()`.
+We included a [test stem file](https://www.heise.de/ct/artikel/c-t-Remix-Wettbewerb-The-Easton-Ellises-2542427.html) (Creative Commons license CC BY-NC-SA 3.0) that can be used by `stempeg.example_stem_path()`.
 
 ### Reading stems
 
@@ -49,33 +51,23 @@ import stempeg
 S, rate = stempeg.read_stems(stempeg.example_stem_path())
 ```
 
-`S` is the stem tensor that includes the time domain signals scaled to `[-1..1]`. The shape is `(stems, samples, channels)`.
+`S` is a numpy tensor that includes the time domain signals scaled to `[-1..1]`. The shape is `(stems, samples, channels)`.
 
-### Reading individual stem ids
+#### Reading individual stem ids
 
-you can read individual substreams of the stem file by passing the corresponding stem id (starting from 0):
+Individual substreams of the stem file can be read by passing the corresponding stem id (starting from 0):
 
 ```python
 S, rate = stempeg.read_stems(stempeg.example_stem_path(), stem_id=[0, 1])
 ```
 
-### Read excerpts (set seek position)
+#### Read excerpts (set seek position)
 
-to read an excerpt from the stem instead of the full file, you can provide start (`start`) and duration (`duration`) in seconds to `read_stems`:
+Excerpts from the stem instead of the full file can be read by providing start (`start`) and duration (`duration`) in seconds to `read_stems`:
 
 ```python
 S, _ = stempeg.read_stems(stempeg.example_stem_path(), start=1, duration=1.5)
 # read from second 1.0 to second 2.5
-```
-
-### Improve performance
-
-if `read_stems` is called repeatedly, it always does two system calls, one for getting the [file info](https://github.com/faroit/stempeg/blob/a56349d2a8297ccf5db13712fc16048029503b26/stempeg/read.py#L120) and one for the [actual reading](https://github.com/faroit/stempeg/blob/a56349d2a8297ccf5db13712fc16048029503b26/stempeg/read.py#L160).  To speed this up you could provide the `Info` object to `read_stems` if the number of streams, the number of channels and the samplerate is identical.
-
-```python
-file_path = stempeg.example_stem_path()
-info = stempeg.Info(file_path)
-S, _ = stempeg.read_stems(file_path, info=info)
 ```
 
 ### Writing stems
@@ -86,13 +78,28 @@ Writing stem files from a numpy tensor
 stempeg.write_stems(S, "output.stem.mp4", rate=44100)
 ```
 
-> :warning: __Warning__: Muxing stems using _ffmpeg_ might lead to non-conform stem files. Please use MP4Box, if you need a reliable result.
-
+> :warning: __Warning__: Muxing stems using _ffmpeg_ leads to multi-stream files not compatible with Native Instrument Hardware or Software. Please use [MP4Box](https://github.com/gpac/gpac) and use the `stempeg.NISTemsWriter()`
 
 ### Use the command line tools
 
 _stempeg_ provides a convenient cli tool to convert a stem to multiple wavfiles. The `-s` switch sets the start, the `-t` switch sets the duration.
 
-
 ```bash
 stem2wav The Easton Ellises - Falcon 69.stem.mp4 -s 1.0 -t 2.5
+```
+
+## F.A.Q
+
+#### How can I improve the reading performance?
+
+`read_stems` is called repeatedly, it always does two system calls, one for getting the file info and one for the actual readingTo speed this up you could provide the `Info` object to `read_stems` if the number of streams, the number of channels and the samplerate is identical.
+
+```python
+file_path = stempeg.example_stem_path()
+info = stempeg.Info(file_path)
+S, _ = stempeg.read_stems(file_path, info=info)
+```
+
+#### How can the quality of the encoded stems be increased
+
+For __Encoding__ it is recommended to use the Fraunhofer AAC encoder (`libfdk_aac`) which is not included in the default ffmpeg builds. Note that the conda version currently does _not_ include `fdk-aac`. If `libfdk_aac` is not installed _stempeg_ will use the default `aac` codec which will result in slightly inferior audio quality.
