+++ stempeg/__init__.py
@@ -72,107 +72,3 @@ def ffmpeg_version():
         return None
 
 
-def cli(inargs=None):
-    """
-    Commandline interface for receiving stem files
-    """
-
-    parser = argparse.ArgumentParser()
-
-    parser.add_argument(
-        '--version', '-V',
-        action='version',
-        version='%%(prog)s %s' % __version__
-    )
-
-    parser.add_argument(
-        'filename',
-        metavar="filename",
-        help="Input STEM file"
-    )
-
-    parser.add_argument(
-        '--format',
-        metavar='format',
-        type=str,
-        default='wav',
-        help="Output format"
-    )
-
-    parser.add_argument(
-        '--id',
-        metavar='id',
-        type=int,
-        nargs='+',
-        help="A list of stem_ids"
-    )
-
-    parser.add_argument(
-        '-s',
-        type=float,
-        nargs='?',
-        help="start offset in seconds"
-    )
-
-    parser.add_argument(
-        '-t',
-        type=float,
-        nargs='?',
-        help="read duration"
-    )
-
-    parser.add_argument(
-        'outdir',
-        metavar='outdir',
-        nargs='?',
-        help="Output folder"
-    )
-
-    args = parser.parse_args(inargs)
-    stem2files(
-        args.filename,
-        args.outdir,
-        args.format,
-        args.id,
-        args.s,
-        args.t
-    )
-
-
-def stem2files(
-    stems_file,
-    outdir=None,
-    format="wav",
-    idx=None,
-    start=None,
-    duration=None,
-):
-    info = Info(stems_file)
-    S, sr = read_stems(stems_file, stem_id=idx, start=start, duration=duration)
-
-    rootpath, filename = op.split(stems_file)
-
-    basename = op.splitext(filename)[0]
-    if ".stem" in basename:
-        basename = basename.split(".stem")[0]
-
-    if outdir is not None:
-        if not op.exists(outdir):
-            os.makedirs(outdir)
-
-        rootpath = outdir
-
-    if len(set(info.title_streams)) == len(info.title_streams):
-        # titles contain duplicates
-        # lets not use the metadata
-        stem_names = info.title_streams
-    else:
-        stem_names = None
-
-    write_stems(
-        op.join(rootpath, basename, "*." + format),
-        data=S,
-        sample_rate=sr,
-        stems_as_files=True,
-        stem_names=stem_names
-    )
